I relicensed and added more sprites.
Yes! I am going back to this game!

Because I read that the Aftero GPL License is not free, I definitely switched back to the regular GPL.

I got out of my rut and drew the human designs of all of the opponents, albeit without their weapons. I also modified the fireball sprite and added a couple of warp sprites.