I added lists of floor tiles.
I made a list of total floor tiles, this list forming the basis of a
list of tiles at where the opponents could be located during setting up
the dungeon and a list of tiles that are not occupied by any of the
opponent's weapons.

I alto made functions that populate these lists.