I started working on the patrol system.
I took note of a new sound effect that I should make.
I rearranged the population function and added a component that has the
target.
I made a system that locates the dragon.
All "is not" conditionals should be the same.
The A* algorithm should now be a proper search that returns an actual
usable queue.
I started working on the patrol, hunt, and fight systems. All of these
have conditionals that determine the opponent's "emotion". Right now, I
found a way of telling the opponent the next target to where the
opponent should go.
I fixed the self-destruct system, though that system is still
incomplete.