I made a system that returns everyone to their places.
I made a system that returns everyone to their places. (This is supposed
to be run after the dragon loses an opportunity because the dragon is
supposed to try again with the same dungion, not generate a new dungeon
and new opponents.)

I also fixed a tiny syntax error with another system.