I trimmed the constants' names.
All of the constants lost their "_constant" suffix because their naming
scheme combined with where and how they are declared made obvious that
they are supposed to be constants anyways.

Also:
I also updated the flag constants to the new naming scheme.
I caught a few out-of-date constant names used in functions that I
forgot to update.