I set the opponents' cross of sight.
I made a new system that determine's the opponent's cross (no longer
"line") of sight.

Also:
I made a current_dungeon_size variable and used that on processes that
requested the local/non-existent dungeon_size variable.
I made a system that switches an opponent from patrolling to hunting.
That system is separated from the fighting system.