I made a "panic mode" system.
I made a system that sets up the panic mode.
I also made a tentative function that should unset the panic mode, but I
am not sure how to code a way of determining that *no entity* that has a
component has that component set to that particular value.

Also:
I reworded the functions in the entity-component system into one that
matches the conventions that the other functions in the code have.