I added the full fight system.
Every opponent has a fight system… or, rather, a "jump to the dragon
then hurt the dragon" system.

Also:

The collision system only inteacted with maps. I tried to expand the
collision system to entities, but the current entity-component system
seems to not know how to interact with two different entities! I did the
best I could with that idea, though.