The game now can play every song.
I made some boilerplate code on some functions that dictatet the state
of the game. At this commit, the code inside these simply say what music
should play.

A consequence of using every song in the code is that the panic music is
used more intelligently inside the hunting system. The separate
panic_system was subsequently redundant and had since been removed.

Also, I renamed the flags of the gameplay phases to normal_phase and
panic_phase.