I got rid of the position and goal components.
The position and goal components depended on the location and target
components. However, because the location component depended on a
function and the target component was an empty table, the position and
goal components got nil values. At the risk of legibility, I replaced
the position and goal components to the parts of the location and target
components to which they originally referred.

Fixes:
I switched an erroneous while loop to a correct for loop.
I renamed the sentinel variables to iterator variables, harmonizing the
names of pure sentinel/iterator variables.
I fixed the scope of random_location in place_subordinate().
I fixed a few spelling errors.
I fixed an erroneous switch of nil to {}.