I fixed collector_of_opponent_setup_cells().
The collector of opponent cells now uses 1 for loop that uses a proper
iterator instead of 2 for loops that use the dungeon's dimmensions, even
though the dimmensions are not used in the algorithm. I also fixed a
reference to the size of opponent_setup_floor_locations that should have
been a reference to the size of total_floor_locations.

Also:
Turns out that odd sizes work with the dungeon-generator in practice. I
adjusted the incrementor and plan_dungeon() in light of this.
I got rid of the debugging code.