I got rid of all table-related syntactic-sugar.
Because the entity-component system is basid on tables (and tables _of_
tables), I decided that getting rid of the syntactic sugar regarding
tables made the structure of the world of the entity-component system
much more clear.

Also:
start_patrolling_system no longer checks if the current phase is the
normal phase (despite the system running during the setup phase).
The correct sprites are drawn in the lance-drawing system.
control_dragon_system no longer checks the sprite component despite that
system not using that component.
I modified the gameplay music a little.